mental_health_chatbot/requirements.txt
openai
langchain
groq
langchain-groq
pandas
numpy
scikit-learn
matplotlib
seaborn
transformers
torch
tensorflow
pillow
scipy
flask
streamlit
huggingface-hub
accelerate
datasets
tiktoken
sentence-transformers
chromadb
langchain-openai
python-dotenv